Caption internal objects w/ tombstone
This is a bit complex because we overwrite the actual object ID in
ArtStories image files with the ID of the ArtStory object. (contextual
images have the wrong id, oops).

We need to run the image filename though an elasticsearch index of image
metadata to obtain the correct ID, then pull the object metadata from
redis, and use that to construct a proper caption.

TODO: the export @damnabit used to populate image metadata in redis is
only for primary images if I'm not mistaken. It hits for most of the
internal images that need captions, but there are a few misses. Can we
pull metadata directly from mediabin for all images?